Added 2010 Normals for the US
Added actual station data so that a few high elevation points could be
represented in the data. Otherwise the WorldClim.org grids are only at a
resolution of 4 km, muting the sharper peaks in elevation. The WorldClim
grids are approximately 1990 normals. Future edits will attempt to
correct for this descrepency with linear regression between the data
sources.