Add matrix to show difference in 1990 and 2010 normals
Mean monthly temperature is linear difference. Low temperature is related to the log of the difference from the mean to ensure that it never gets warmer than mean.  Precipitation is related to the log since the distribution across the area is not linear, and some sites could fall below zero if related to linear differences.